+++ lexyacc-code/y.tab.c
@@ -72,7 +72,7 @@
 /* prototypes */
 nodeType *opr(int oper, int nops, ...);
 nodeType *id(int i);
-nodeType *con(int value);
+nodeType *con(long long value);
 void freeNode(nodeType *p);
 int ex(nodeType *p);
 int yylex(void);
@@ -156,7 +156,7 @@ union YYSTYPE
 {
 #line 19 "calc3.y" /* yacc.c:355  */
 
-    int iValue;                 /* integer value */
+    long long iValue;                 /* integer value */
     char sIndex;                /* symbol table index */
     nodeType *nPtr;             /* node pointer */
 
@@ -1719,7 +1719,7 @@ yyreturn:
 
 #define SIZEOF_NODETYPE ((char *)&p->con - (char *)p)
 
-nodeType *con(int value) {
+nodeType *con(long long value) {
     nodeType *p;
     size_t nodeSize;
